Fix missing modules and server path
